Move getConfig to utils, add config values to process.env
