+++ src/index.js
@@ -1,6 +1,5 @@
 const path = require('path')
 const fs = require('fs-extra')
-const { defaultsDeep } = require('lodash')
 const chokidar = require('chokidar');
 
 const {
@@ -10,9 +9,19 @@ const {
   symlinkExtension,
   copyDependencies,
   copyIcons,
+  getConfig,
+  objectToProcessEnv,
 } = require('./utils')
 
 module.exports = async bundler => {
+  const root = process.cwd()
+  // load package.json
+  const package = fs.readJsonSync(path.join(root, 'package.json'))
+  // load config
+  const config = getConfig(package)
+  // assign config values to process.env
+  objectToProcessEnv(config)
+  // only run when the process is the one bundling the .html file
   if (bundler.entryFiles.length === 1 && path.extname(bundler.entryFiles[0]) === '.html') {
     const htmlFilename = path.basename(bundler.entryFiles[0])
     const env = process.env.NODE_ENV
@@ -20,9 +29,8 @@ module.exports = async bundler => {
       env != 'production' && bundler.server
         ? bundler.server.address().port
         : 1234
-    const root = process.cwd()
     const out = bundler.options.outDir
-    const package = fs.readJsonSync(path.join(root, 'package.json'))
+    // listen for changes to the package.json (that might have gotten changes to cep config values) and re-bundle
     const watch = chokidar.watch(path.join(root, 'package.json'), {
       ignored: /(^|[\/\\])\../,
       persistent: true
@@ -32,43 +40,6 @@ module.exports = async bundler => {
     })
     bundle()
     async function bundle() {
-      const config = defaultsDeep(
-        {
-          bundleName: process.env.NAME,
-          bundleId: process.env.ID,
-          bundleVersion: process.env.VERSION,
-          hosts: process.env.HOSTS,
-          iconNormal: process.env.ICON_NORMAL,
-          iconRollover: process.env.ICON_ROLLOVER,
-          iconDarkNormal: process.env.ICON_DARK_NORMAL,
-          iconDarkRollover: process.env.ICON_DARK_ROLLOVER,
-          panelWidth: process.env.PANEL_WIDTH,
-          panelHeight: process.env.PANEL_HEIGHT,
-        },
-        {
-          bundleName: package.cep && package.cep.name,
-          bundleId: package.cep && package.cep.id,
-          bundleVersion: package.cep && package.cep.version,
-          hosts: package.cep && package.cep.hosts,
-          iconNormal: package.cep.iconNormal,
-          iconRollover: package.cep.iconRollover,
-          iconDarkNormal: package.cep.iconDarkNormal,
-          iconDarkRollover: package.cep.iconDarkRollover,
-          panelWidth: package.cep.panelWidth,
-          panelHeight: package.cep.panelHeight,
-        },
-        {
-          bundleVersion: package.version,
-        },
-        {
-          bundleName: 'My Extension',
-          bundleId: 'com.mycompany.myextension',
-          bundleVersion: '0.0.1',
-          hosts: '*',
-          panelWidth: 500,
-          panelHeight: 500,
-        }
-      )
       enablePlayerDebugMode()
       const hosts = parseHosts(config.hosts)
       await copyDependencies({ root, out, package })
@@ -92,6 +63,4 @@ module.exports = async bundler => {
       await copyIcons({ bundler, config })
     }
   }
-
-  // this.watchedDirectories = new Map();
 }
