+++ src/utils.js
@@ -3,6 +3,11 @@ const path = require('path')
 const fs = require('fs-extra')
 const os = require('os')
 const { range } = require('lodash')
+const { defaultsDeep } = require('lodash')
+
+const manifestTemplate = require('./templates/manifest')
+const panelTemplate = require('./templates/panel')
+const debugTemplate = require('./templates/.debug')
 
 function templateDebug(formatter) {
   return range(4, 16)
@@ -41,9 +46,84 @@ function disablePlayerDebugMode() {
   }
 }
 
-const manifestTemplate = require('./templates/manifest')
-const panelTemplate = require('./templates/panel')
-const debugTemplate = require('./templates/.debug')
+function camelToSnake(str) {
+  return str.replace(/([A-Z])/g, (part) => `_${part.toLowerCase()}`)
+}
+
+function getConfig(package) {
+  const debugPortEnvs = Object.keys(process.env)
+    .filter((key) => key.indexOf('DEBUG_PORT_') === 0)
+  const config = defaultsDeep(
+    {
+      bundleName: process.env.NAME,
+      bundleId: process.env.ID,
+      bundleVersion: process.env.VERSION,
+      hosts: process.env.HOSTS,
+      iconNormal: process.env.ICON_NORMAL,
+      iconRollover: process.env.ICON_ROLLOVER,
+      iconDarkNormal: process.env.ICON_DARK_NORMAL,
+      iconDarkRollover: process.env.ICON_DARK_ROLLOVER,
+      panelWidth: process.env.PANEL_WIDTH,
+      panelHeight: process.env.PANEL_HEIGHT,
+      debugPorts: debugPortEnvs.length > 0
+        ? debugPortEnvs.reduce((obj, key) => {
+          obj[key] = process.env[key];
+          return obj
+        }, {})
+        : undefined
+    },
+    {
+      bundleName: package.cep && package.cep.name,
+      bundleId: package.cep && package.cep.id,
+      bundleVersion: package.cep && package.cep.version,
+      hosts: package.cep && package.cep.hosts,
+      iconNormal: package.cep.iconNormal,
+      iconRollover: package.cep.iconRollover,
+      iconDarkNormal: package.cep.iconDarkNormal,
+      iconDarkRollover: package.cep.iconDarkRollover,
+      panelWidth: package.cep.panelWidth,
+      panelHeight: package.cep.panelHeight,
+      debugPorts: package.cep.debugPorts,
+    },
+    {
+      bundleVersion: package.version,
+    },
+    {
+      bundleName: 'Parcel CEP Extension',
+      bundleId: 'com.mycompany.myextension',
+      bundleVersion: '0.0.1',
+      hosts: '*',
+      panelWidth: 500,
+      panelHeight: 500,
+      debugPorts: {
+        PHXS: 3001,
+        IDSN: 3002,
+        AICY: 3003,
+        ILST: 3004,
+        PPRO: 3005,
+        PRLD: 3006,
+        AEFT: 3007,
+        FLPR: 3008,
+        AUDT: 3009,
+        DRWV: 3010,
+        MUST: 3011,
+        KBRG: 3012,
+      }
+    }
+  )
+  return config
+}
+
+function objectToProcessEnv(object) {
+  // assign object to process.env so they can be used in the code
+  Object.keys(object).forEach(key => {
+    const envKey = camelToSnake(key).toUpperCase()
+    const value = typeof object[key] === 'string'
+      ? object[key]
+      : JSON.stringify(object[key])
+    process.env[envKey] = value
+  })
+}
 
 async function writeExtensionTemplates({
   env,
@@ -105,7 +185,6 @@ function parseHosts(hostsString) {
       } else if (version) {
         version = version
       }
-
       return {
         name,
         version,
@@ -195,4 +274,6 @@ module.exports = {
   writeExtensionTemplates,
   parseHosts,
   symlinkExtension,
+  getConfig,
+  objectToProcessEnv,
 }
